Подключил Hibernate. Пока только чтоб запускалось

+++ SimpleServiceHibernate/src/main/java/service/CourseService.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public interface CourseService {
 
-    Course findById(Long id);
+    Course findById(Integer id);
 
     void save(Course course);
 
     void update(Course course);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     List<Course> findAll();
